- Replaced placement allocator by to step heap allocator with early and normal init - Shortened _initialized_get functions to _init_get - Added early heap init call at beginning of kernel_main - Reworked heap to be capable to hold two stages - Bound shrink and expand heap to normal heap - Changed framebuffer_putc call to check also for framebuffer initialized - Added initial heap of 1MB at linker script - Fixed logical issue within physical bitmap allocator - Moved event init, interrupt init, interrupt enable and gdb init to beginning - Replaced max memory used within early boot by using __kernel_end - Removed obsolete function boot_platform_prepare

+++ src/platform/rpi/mailbox/property.c
@@ -25,7 +25,6 @@
 #include <core/debug/debug.h>
 #include <core/entry.h>
 #include <core/mm/phys.h>
-#include <core/mm/placement.h>
 #include <platform/rpi/mailbox/mailbox.h>
 #include <platform/rpi/mailbox/property.h>
 
@@ -47,7 +46,7 @@ void mailbox_property_init( void ) {
   // reserve memory if not yet done
   if ( NULL == ptb_buffer ) {
     ptb_buffer = ( int32_t* )aligned_alloc( PAGE_SIZE, PAGE_SIZE );
-    ptb_buffer_phys = ( int32_t* )VIRT_2_PHYS( ( uintptr_t )ptb_buffer );
+    ptb_buffer_phys = ( int32_t* )VIRT_2_PHYS( ptb_buffer );
   }
 
   // Add startup size
